requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 6
cellpose == 2.1.0

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 9
docker[tls]~=5.0.3

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 8
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 5
h5py == 3.7.0

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 9,10
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 8,18
holoviews == 1.15.0

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 16
hvplot == 0.8.0

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 5,6
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 2
numpy == 1.22.0


# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 14
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 13
pandas == 1.4.3

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\analysis_app.py: 4
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 3
panel == 0.13.1

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\analysis_app.py: 3
# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 4
param == 1.12.2

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 7
pillow == 9.2.0

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 14
scipy == 1.8.1

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 15
shapely == 1.8.2

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\ui.py: 12
scikit-image[optional] ==  0.19.3

# C:\Users\leo\PycharmProjects\MicroscopyPipeline\utils.py: 17
tqdm == 4.64.0

aicsimageio == 3.2.1
scikit-learn == 1.1.1
#torch~=1.12.0
typer~=0.5.0
PyYAML~=6.0
setuptools~=60.2.0
pydantic~=1.9.1
tifffile~=2022.5.4
psutil~=5.9.1
torch~=1.12.0+cu116